Create ETH Base
A contract featuring a single wallet owner, which can receive funds with a fallback function, spend funds on both EOAs and contracts, and delegate allowances to other addresses. Additionally, it includes the capability to assign a new owner with the approval of 3 out of 5 designated guardians. This 3-of-5 multi-signature mechanism ensures that, in case of the owner losing access to the wallet, a new user can gain access if 3-5 guardians agree to transfer access control.